+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 [unreleased]
 
 - Added support for npy version 2.0
+- Fixed error with empty shapes
 
 ## 0.1.0